+++ src/idpconfgen/cli_merge.py
@@ -7,20 +7,23 @@ copy and re-number them to make 500 conformers in the working dir. or
 destination dir.
 
 USAGE:
-    $ idpconfgen merge --target <PATH_TO_SUBFOLDERS>
+    $ idpconfgen merge --source <PATH_TO_SUBFOLDERS>
     $ idpconfgen merge \\
-        --target <PATH_TO_SUBFOLDERS> \\
+        --source <PATH_TO_SUBFOLDERS> \\
         --destination <PATH_TO_OUTPUT> \\
         --prefix <CUSTOM_NAME> \\
-        --delete \\
+        --delete
 """
+import argparse
+import glob
+import shutil
+from pathlib import Path
 
-import argparse, shutil, glob
-
-from idpconfgen.libs.libio import make_folder_or_cwd
-from idpconfgen.libs import libcli
 from idpconfgen import log
+from idpconfgen.libs import libcli
+from idpconfgen.libs.libio import make_folder_or_cwd
 from idpconfgen.logger import S, T, init_files
+
 
 LOGFILESNAME = 'idpconfgen_merge'
 _name = 'merge'
@@ -36,90 +39,98 @@ ap = libcli.CustomParser(
     )
 
 ap.add_argument(
-    '-tgt',
-    '--target',
-    help=('Directory containing subfolders of conformers to merge. Required.'
-          ),
-    type=str,
+    '-src',
+    '--source',
+    help=(
+        'Directory containing subfolders of conformers to merge. '
+        'Defaults to the current working directory.'
+        ),
+    type=Path,
     required = True,
-    default='./',
-)
+    default=None,
+    )
 
 ap.add_argument(
     '-des',
     '--destination',
-    help=('Directory to collate conformers. Optional.'
-          ),
-    type=str,
+    help=(
+        'Directory to collate conformers. Optional. '
+        'Defaults to the current working directory.'
+        ),
+    type=Path,
     default=None,
-)
+    )
 
 ap.add_argument(
     '-pre',
     '--prefix',
-    help=('Name to replace default `conformer` for each file. Optional.'
-          ),
+    help='Common prefix of the new files. Defaults to `conformer`.',
     type=str,
     default='conformer',
-)
+    )
 
 ap.add_argument(
     '-del',
     '--delete',
-    help='Deleted conformers and subfolders after merging. Defaults to True.',
+    help='Deleted conformers and subfolders after merging. Defaults to False.',
     action='store_true',
-)
+    )
+
 
 def main(
-        target,
+        source,
         destination=None,
         prefix='conformer',
         delete=False,
         **kwargs,
-    ):
+        ):
     """
     Perform main logic of the script.
 
     Parameters
     ----------
-    target : str, required
+    source : str, required
         Path to the directory with all of the subfolders.
-    
+
     destination : str, optional
-        Path to the directory to save the collated conformers from the target.
-    
+        Path to the directory to save the collated conformers from the source.
+
     prefix : str, optional
         Replace the default `conformer` name for each file.
-    
+
     delete : bool, optional
         Flag to turn on deletion of conformers and subfolders of interest.
         Defaults to True.
     """
     init_files(log, LOGFILESNAME)
-    
+
     # initialize some variables
     count=0
-    if destination == None: destination = target
-    destination = str(make_folder_or_cwd(destination))
-    
+
+    if destination is None:
+        destination = source
+
+    source = make_folder_or_cwd(source)
+    destination = make_folder_or_cwd(destination)
+
     log.info(T('Attempting to collate files'))
     log.info(S('Obtaining paths to all conformers...'))
     # get all the paths to the subfolders
-    subpaths = list(glob.glob(f'{target}/*/'))
-    pdbpaths = list(glob.glob(f'{target}/*/*.pdb', recursive=True))
-    
+    subpaths = list(glob.glob(f'{source}/*/'))
+    pdbpaths = list(glob.glob(f'{source}/*/*.pdb', recursive=True))
+
     log.info(S('done'))
-    
+
     log.info(S(f'Copying {len(pdbpaths)} files with prefix {prefix}...'))
     for path in pdbpaths:
         # using copy2 to preserve metadata and file permissions
         try:
-            shutil.copy2(path, destination+f'/{prefix}_{count}.pdb')
+            shutil.copy2(path, Path(destination, f'{prefix}_{count}.pdb'))
         except OSError as e:
             log.info(S(f'Error: {path} : {e.strerror}'))
         count+=1
     log.info(S('done'))
-    
+
     if delete:
         log.info(S('Cleaning up subfolders and PDB files...'))
         for path in subpaths:
@@ -128,5 +139,5 @@ def main(
             except OSError as e:
                 log.info(S(f'Error: {path} : {e.strerror}'))
         log.info(S('done'))
-    
+
     return